requirements.txt
cryptography==37.0.2
fastapi==0.78.0
fastapi-mail==1.1.4
PyJWT==2.4.0
pymongo==4.1.1